% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{api_key}
\alias{api_key}
\title{Get the api key for the header}
\usage{
api_key(key = NA)
}
\arguments{
\item{key}{Users api key, Default: NA}
}
\value{
Character string of the users api key.
}
\description{
either use the one that is supplied, or find it from the .Renviron
}
\details{
Will return the character string given, if given. If none is supplied, will look for it as environmental variable.
}
\examples{
\dontrun{
if(interactive()){
  api_key
 }
}
}
