% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_article.R
\name{create_new_article}
\alias{create_new_article}
\title{Create a new article locally}
\usage{
create_new_article(title, series = "series",
  tags = "[\\"tag1\\", \\"tag2\\"]", file = "")
}
\arguments{
\item{title}{string of the title}

\item{series}{string of the series, Default: 'series'}

\item{tags}{string of the YAML array of tags, Default: '['tag1', 'tag2']'}

\item{file}{path of the file to create, should be a "*.Rmd", Default: ''}
}
\value{
By default the yaml boilerplate to the command line. If a file path is included it will create that file.
}
\description{
Make a new file with boilerplate YAML for posting to dev.to
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 create_new_article(title = "my title")
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
