% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_me.R
\name{get_me}
\alias{get_me}
\title{get my articles}
\usage{
get_me(key = NA)
}
\arguments{
\item{key}{the api you have set up on DEV.TO}
}
\value{
article stuff
}
\description{
Provides lots of info on your users articles
}
\details{
if no key is supplied, will check for key named DEVTO in `.Renviron`
}
\examples{
\dontrun{
if(interactive()){
 get_me("my_api_key")
 }
}
}
\seealso{
\code{\link[httr]{content}},\code{\link[httr]{GET}},\code{\link[httr]{add_headers}}
}
