% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_user.R
\name{get_my_user}
\alias{get_my_user}
\title{Get the authenticated user}
\usage{
get_my_user(key = NA, tidy = TRUE)
}
\arguments{
\item{key}{the api you have set up on DEV.TO, Default: NA}

\item{tidy}{if the response should be parsed into a tibble, Default: TRUE}
}
\value{
user stuff
}
\description{
Provides information on the authenticated user
}
\details{
if no key is supplied, will check for key named DEVTO in `.Renviron`
}
\examples{
\dontrun{
if(interactive()){
 get_my_user("my_api_key")
 }
}
}
\seealso{
\code{\link[httr]{content}},\code{\link[httr]{GET}},\code{\link[httr]{add_headers}}
 \code{\link[dev.to.ol]{api_key}}
}
