% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubify_images.R
\name{githubify_images}
\alias{githubify_images}
\title{FUNCTION_TITLE}
\usage{
githubify_images(
  article_string,
  pattern = "/home/davidparr/Dev/dev.to-posts/",
  user = "DaveParr",
  repo = "dev.to-posts",
  branch = "master"
)
}
\arguments{
\item{article_string}{PARAM_DESCRIPTION}

\item{pattern}{PARAM_DESCRIPTION, Default: '/home/davidparr/Dev/dev.to-posts/'}

\item{user}{PARAM_DESCRIPTION, Default: 'daveparr'}

\item{repo}{PARAM_DESCRIPTION, Default: 'dev.to-posts'}

\item{branch}{PARAM_DESCRIPTION, Default: 'master'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readr]{read_file}}
 \code{\link[glue]{glue}}
}
