% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_postable_Rmd.R
\name{is_postable_Rmd}
\alias{is_postable_Rmd}
\title{Check a file is postable article}
\usage{
is_postable_Rmd(file)
}
\arguments{
\item{file}{file path}
}
\value{
TRUE, or FALSE with message
}
\description{
Will check file is readable and extention is .Rmd
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 is_postable_Rmd("./my/file.Rmd)
 attr(is_postable_Rmd("./my/file.Rmd), "msg")
 }
}
}
\seealso{
\code{\link[assertthat]{assert_that}},\code{\link[assertthat]{assertions-file}}
}
