% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_article.R
\name{parse_article}
\alias{parse_article}
\title{prepare an article for publication on dev.to}
\usage{
parse_article(file)
}
\arguments{
\item{file}{.Rmd file}
}
\value{
list
}
\description{
Uses the YAML frontmatter of an .Rmd to create a raw markdown file suitable for publication.
}
\details{
The list is describing the article metadata, markdown article output location and markdown article string. The title is removed from the article string and is present in the metadata.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rmarkdown]{yaml_front_matter}},\code{\link[rmarkdown]{render}}
 \code{\link[purrr]{map}}
 \code{\link[stringr]{str_remove}}
 \code{\link[readr]{read_file}}
 \code{\link[glue]{glue}}
}
