% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_new_article.R
\name{post_new_article}
\alias{post_new_article}
\title{Post a markdown file to dev.to}
\usage{
post_new_article(file, key = NA)
}
\arguments{
\item{file}{The path to the file}

\item{key}{Your API key, Default: NA}
}
\value{
The response
}
\description{
Create a new post from an .Rmd.
}
\details{
Will look for an api key in the `.REnviron` file. Seems to check if the body is identical to a previous article and error if so with `"Body markdown has already been taken"`.
The following YAML arguments are read from the file YAML frontmatter if present:
\describe{
  \item{title}{A character string}
  \item{series}{A character string}
  \item{published}{A boolean}
  \item{tags}{list of character strings: \code{["tag1", "tag2"]}}
}

dev.to will not allow tags with spaces, so tags with spaces will be automatically collapsed.

The default table output method renders a very large print code block.
The workaround is to use  \code{\link[knitr]{kable}}.
}
\examples{
\dontrun{
if(interactive()){
 post_new_article("./articles/my_article.Rmd")
 }
}
}
\seealso{
\code{\link[rmarkdown]{yaml_front_matter}},\code{\link[rmarkdown]{render}}
 \code{\link[readr]{read_file}}
 \code{\link[stringr]{str_remove}}
 \code{\link[glue]{glue}}
 \code{\link[httr]{POST}},\code{\link[httr]{add_headers}},\code{\link[httr]{content}}
}
