% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_environ.R
\name{setup_environ}
\alias{setup_environ}
\title{Setup .Renviron}
\usage{
setup_environ(
  devto,
  github_user,
  github_repo,
  github_branch = "main",
  local_image_path = getwd()
)
}
\arguments{
\item{devto}{Your API key}

\item{github_user}{Your GitHub Username}

\item{github_repo}{Your GitHub Repo with your articles}

\item{github_branch}{Your GitHub branch that you are publishing from, Default:'main'}

\item{local_image_path}{The local path prefix your images get stored at, Default: getwd()}
}
\value{
A brand new .Renviron file
}
\description{
Write a .Renviron file with the details needed for dev.to
}
\details{
To embed the links to images in dev.to posts, dev.to.ol uses GitHub
as an image hosting service. This means that your locally generated images (like graphs)
should be committed to a public repo. \code{\link[dev.to.ol]{githubify_images}}
does the work of identifying a local image link and replacing it with the expected
location on GitHub.
}
\examples{
\dontrun{
if(interactive()){
 setup_environ(devto = "mysecretkey", github_user = "MyName", github_repo = "dev.to-posts")
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
