% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_article.R
\name{update_article}
\alias{update_article}
\title{Update a post on dev.to}
\usage{
update_article(id, file, key = NA)
}
\arguments{
\item{id}{The id of the post to update as a string}

\item{file}{The path to the file}

\item{key}{Your API key, Default: NA}
}
\value{
A response object
}
\description{
Updates an existing post on dev.to by id with content from and .Rmd
}
\details{
The id of the post must be supplied, and can be returned by the \code{\link[dev.to.ol]{get_users_articles}} function
}
\examples{
\dontrun{
if(interactive()){
 update_article(id = "369520", file = "./article.Rmd")
 }
}
}
\seealso{
\code{\link[rmarkdown]{yaml_front_matter}},\code{\link[rmarkdown]{render}}
 \code{\link[purrr]{map}}
 \code{\link[stringr]{str_remove}}
 \code{\link[readr]{read_file}}
 \code{\link[glue]{glue}}
 \code{\link[httr]{PUT}},\code{\link[httr]{add_headers}}
}
