% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_articles.R
\name{get_users_articles}
\alias{get_users_articles}
\title{Get the authenticated users articles}
\usage{
get_users_articles(key = NA, tidy = TRUE)
}
\arguments{
\item{key}{the api key you have set up on DEV.TO, Default: NA}

\item{tidy}{if the response should be parsed into a tibble, Default: TRUE}
}
\value{
article stuff
}
\description{
Provides lots of info on your users articles
}
\details{
If no \code{key} is supplied, will check for key named DEVTO in `.Renviron`. If \code{tidy} is \code{FALSE}, then the full, raw response is returned. Returns both published and unpublished articles, in 2 consecutive calls.
}
\examples{
\dontrun{
if(interactive()){
 get_users_articles("my_api_key")
 }
}
}
\seealso{
\code{\link[httr]{GET}},\code{\link[httr]{add_headers}},\code{\link[httr]{content}}
 \code{\link[tibble]{tibble}}
 \code{\link[tidyr]{hoist}}
 \code{\link[dplyr]{bind}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}}
}
