% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubify_images.R
\name{githubify_images}
\alias{githubify_images}
\title{Replace local links to files with the expected github location.}
\usage{
githubify_images(
  article_string,
  local_image_path = Sys.getenv(x = "LOCAL_IMAGE_PATH"),
  github_user = Sys.getenv(x = "GITHUB_USER"),
  github_repo = Sys.getenv(x = "GITHUB_REPO"),
  github_branch = Sys.getenv(x = "GITHUB_BRANCH")
)
}
\arguments{
\item{article_string}{The article as a string, e.g. from `readr::read_file`}

\item{local_image_path}{The local path to the images, Default: Sys.getenv(x = "LOCAL_IMAGE_PATH")}

\item{github_user}{Your github username, Default: Sys.getenv(x = "GITHUB_USER")}

\item{github_repo}{The github repo hosting the article and image, Default: Sys.getenv(x = "GITHUB_REPO")}

\item{github_branch}{The branch on the github repo to publish from, Default: Sys.getenv(x = "GITHUB_BRANCH")}
}
\value{
The article string with the image paths changed from a local path to an expected location on GitHub.
}
\description{
The function will look for capitalised environmental variables of the same name as the argument set in an .Renviron
}
\details{
At no point does dev.to.ol check that the image is at the expected path, so remember to push your work,
and double check the paths are set correctly. \code{\link[dev.to.ol]{setup_environ}} can help with this.
}
\seealso{
\code{\link[readr]{read_file}}
 \code{\link[glue]{glue}}
}
