% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_article.R
\name{parse_article}
\alias{parse_article}
\title{FUNCTION_TITLE}
\usage{
parse_article(file)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rmarkdown]{yaml_front_matter}},\code{\link[rmarkdown]{render}}
 \code{\link[purrr]{map}}
 \code{\link[stringr]{str_remove}}
 \code{\link[readr]{read_file}}
 \code{\link[glue]{glue}}
}
