% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_new_article.R
\name{post_new_article}
\alias{post_new_article}
\title{Post a markdown file to dev.to}
\usage{
post_new_article(file, series = NA, tags = NA, key = NA)
}
\arguments{
\item{file}{The path to the file, Default: file}

\item{series}{Character string of the series name, Default: NA}

\item{tags}{List of tags, Default: NA}

\item{key}{Your API key, Default: NA}
}
\value{
The response
}
\description{
Create a new post from an .Rmd
}
\details{
Will look for an api key in the `.REnviron` file. Seems to check if the body is identical to a previous article and error if so with `"Body markdown has already been taken"`.
}
\examples{
\dontrun{
if(interactive()){
 post_new_article("./articles/my_article.Rmd", tags = list('montag', 'dienstag'))
 }
}
}
\seealso{
\code{\link[httr]{POST}},\code{\link[httr]{add_headers}},\code{\link[httr]{verbose}},\code{\link[httr]{content}}
 \code{\link[readr]{read_file}}, \code{\link[dev.to.ol]{api_key}}, \code{\link[dev.to.ol]{is_postable_article}}
}
